misc/familyGPT/src/requirements.txt
openai==0.27.0
Flask==2.1.2
python-dotenv==1.0.0